Refactor switch-case statement
