+++ app/components/Views/Quiz/QuizContent/QuizContent.tsx
