+++ app/components/Views/Quiz/QuizContent/index.ts
