+++ app/components/Views/Quiz/QuizContent/styles.ts
