+++ app/components/Views/Quiz/SRPQuiz/styles.ts
