+++ app/components/Views/Quiz/index.ts
@@ -1,0 +1,4 @@
+/* eslint-disable import/prefer-default-export */
+import { SRPQuiz } from './SRPQuiz';
+
+export { SRPQuiz };
