--- app/components/Views/SRPQuiz/index.ts
@@ -1,4 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-import { QuizModal } from './QuizModal';
-
-export { QuizModal };
